fix: 습관 통계 LazyInitializationException 해결
